requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
pdfplumber==0.10.3
Pillow==10.1.0
